Fix Audio Rendering
Browsers nowadays won't let you construct an audio renderer except in
response to an event, so this moves the construction of the renderer
into the keypress handler.  This is a little bit of a nasty hack and
it makes Flow unhappy, but it works in practice because nothing needs
to render until after a keypress starts the game loop and I shut Flow
up by changing the renderer's type to any.